+++ src/input.rs
@@ -11,7 +11,6 @@
 use hashbrown::{HashMap, HashSet};
 
 use crate::glm::Vec2;
-use crate::graphics;
 use crate::platform;
 use crate::Context;
 
@@ -412,31 +411,14 @@ pub fn is_mouse_button_released(ctx: &Context, button: MouseButton) -> bool {
 ///
 /// If the screen is scaled, the returned value will be relative to the original size.
 pub fn get_mouse_x(ctx: &Context) -> f32 {
-    let raw_position = ctx.input.mouse_position.x;
-
-    match graphics::get_screen_rect(ctx) {
-        None => raw_position,
-        Some(screen_rect) => {
-            let internal_width = graphics::get_internal_width(ctx) as f32;
-            ((raw_position - screen_rect.x) / screen_rect.width) * internal_width
-        }
-    }
+    ctx.input.mouse_position.x
 }
 
 /// Get the Y co-ordinate of the mouse.
 ///
 /// If the screen is scaled, the returned value will be relative to the original size.
 pub fn get_mouse_y(ctx: &Context) -> f32 {
-    let raw_position = ctx.input.mouse_position.y;
-
-    match graphics::get_screen_rect(ctx) {
-        None => raw_position,
-        Some(screen_rect) => {
-            let internal_height = graphics::get_internal_height(ctx) as f32;
-
-            ((ctx.input.mouse_position.y - screen_rect.y) / screen_rect.height) * internal_height
-        }
-    }
+    ctx.input.mouse_position.y
 }
 
 /// Get the position of the mouse.