+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project adheres to Semantic Versioning.
 * `Key` and `MouseButton` are now Tetra-specific types, rather than re-exporting the SDL versions. Note that some names have been changed for consistency, and some variants have been removed to simplify the docs.
 * `TetraError::Sdl` is now called `TetraError::Platform`.
 * `DEFAULT_VERTEX_SHADER` and `DEFAULT_FRAGMENT_SHADER` are now const instead of static.
-* The renderer now draws directly to the window when `ScreenScaling::Resize` is enabled, avoiding unnecessary framebuffer overhead. 
+* Screen scaling has been extracted from the core of the engine, and is now provided via the `ScreenScaler` struct. This allows it to be more flexibly integrated with the rest of your game's rendering.
 * Updated `glow` to 0.2.3.
 * Updated `hashbrown` to 0.6.0.
 * Updated `image` to 0.22.1.